+++ WASD.QLicPlatform.API/Alerts/Domain/Model/Queries/GetAlertByIdQuery.cs
@@ -1,3 +1,3 @@
 namespace WASD.QLicPlatform.API.Alerts.Domain.Model.Queries;
 
-public record GetAlertByIdQuery(long Id);
+public record GetAlertByIdQuery(int Id);
